pyproject.toml
[project]
name = "hud-ui-cube"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "hud-python>=0.5.0",
  "pyautogui",
  "playwright",
  "httpx",
  "typer",
  "google-api-python-client",
  "google-auth",
  "pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["providers", "tools", "setup", "scenarios"]
